1451012 김한수

+++ src/main/java/next/controller/qna/DeleteAnswerController.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import core.mvc.Controller;
 import next.dao.AnswerDao;
+import next.model.Answer;
 import next.model.Result;
 
 public class DeleteAnswerController implements Controller {
@@ -18,6 +19,12 @@ public class DeleteAnswerController implements Controller {
 		Long answerId= Long.parseLong(req.getParameter("answerId"));
 		
 		AnswerDao answerDao =new AnswerDao();
+		Answer answer = answerDao.findById(answerId);
+		answerDao.delete(answerId);
+		
+		if(answer ==null)
+			return null;
+		
 		answerDao.delete(answerId);
 		
 		ObjectMapper mapper = new ObjectMapper();
